# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DutyBatchCreate": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/DutyCreate"
            },
            "maxItems": 31,
            "minItems": 1,
            "title": "Entries",
            "type": "array"
          }
        },
        "required": [
          "entries"
        ],
        "title": "DutyBatchCreate",
        "type": "object"
      },
      "DutyCreate": {
        "properties": {
          "branch_id": {
            "title": "Branch Id",
            "type": "string"
          },
          "city": {
            "maxLength": 120,
            "minLength": 1,
            "title": "City",
            "type": "string"
          },
          "duty_date": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Duty Date",
            "type": "string"
          },
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "pharmacy_id": {
            "title": "Pharmacy Id",
            "type": "string"
          },
          "source": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Source",
            "type": "string"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          }
        },
        "required": [
          "pharmacy_id",
          "branch_id",
          "city",
          "duty_date",
          "start_at",
          "end_at",
          "source"
        ],
        "title": "DutyCreate",
        "type": "object"
      },
      "DutyOut": {
        "properties": {
          "city": {
            "title": "City",
            "type": "string"
          },
          "duty_date": {
            "title": "Duty Date",
            "type": "string"
          },
          "duty_end_at": {
            "format": "date-time",
            "title": "Duty End At",
            "type": "string"
          },
          "duty_start_at": {
            "format": "date-time",
            "title": "Duty Start At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "verification_status": {
            "title": "Verification Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "city",
          "duty_date",
          "duty_start_at",
          "duty_end_at",
          "status",
          "verification_status"
        ],
        "title": "DutyOut",
        "type": "object"
      },
      "DutyUpdate": {
        "properties": {
          "branch_id": {
            "title": "Branch Id",
            "type": "string"
          },
          "city": {
            "maxLength": 120,
            "minLength": 1,
            "title": "City",
            "type": "string"
          },
          "duty_date": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Duty Date",
            "type": "string"
          },
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "source": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Source",
            "type": "string"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          }
        },
        "required": [
          "branch_id",
          "city",
          "duty_date",
          "start_at",
          "end_at",
          "source"
        ],
        "title": "DutyUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "NaviMed API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/notifications": {
      "get": {
        "operationId": "list_notifications_v1_notifications_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/device": {
      "post": {
        "operationId": "register_device_v1_notifications_device_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register Device",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/{notification_id}/read": {
      "post": {
        "operationId": "mark_read_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "title": "Notification Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark Read",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/v1/pharmacies/{pharmacy_id}/duty-schedules": {
      "get": {
        "operationId": "list_duty_v1_pharmacies__pharmacy_id__duty_schedules_get",
        "parameters": [
          {
            "in": "path",
            "name": "pharmacy_id",
            "required": true,
            "schema": {
              "title": "Pharmacy Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DutyOut"
                  },
                  "title": "Response List Duty V1 Pharmacies  Pharmacy Id  Duty Schedules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Duty",
        "tags": [
          "pharmacy"
        ]
      },
      "post": {
        "operationId": "create_duty_v1_pharmacies__pharmacy_id__duty_schedules_post",
        "parameters": [
          {
            "in": "path",
            "name": "pharmacy_id",
            "required": true,
            "schema": {
              "title": "Pharmacy Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DutyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DutyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Duty",
        "tags": [
          "pharmacy"
        ]
      }
    },
    "/v1/pharmacies/{pharmacy_id}/duty-schedules/batch": {
      "post": {
        "operationId": "create_duty_batch_v1_pharmacies__pharmacy_id__duty_schedules_batch_post",
        "parameters": [
          {
            "in": "path",
            "name": "pharmacy_id",
            "required": true,
            "schema": {
              "title": "Pharmacy Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DutyBatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DutyOut"
                  },
                  "title": "Response Create Duty Batch V1 Pharmacies  Pharmacy Id  Duty Schedules Batch Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Duty Batch",
        "tags": [
          "pharmacy"
        ]
      }
    },
    "/v1/pharmacies/{pharmacy_id}/duty-schedules/{schedule_id}": {
      "put": {
        "operationId": "update_duty_v1_pharmacies__pharmacy_id__duty_schedules__schedule_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "pharmacy_id",
            "required": true,
            "schema": {
              "title": "Pharmacy Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DutyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DutyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Duty",
        "tags": [
          "pharmacy"
        ]
      }
    },
    "/v1/pharmacies/{pharmacy_id}/duty-schedules/{schedule_id}/submit": {
      "post": {
        "operationId": "submit_duty_v1_pharmacies__pharmacy_id__duty_schedules__schedule_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "pharmacy_id",
            "required": true,
            "schema": {
              "title": "Pharmacy Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DutyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Duty",
        "tags": [
          "pharmacy"
        ]
      }
    },
...
  }
}